annual analysis table shared for absolute consumption and regression
