+++ src/app/analysis/calculations/analysis-calculations-helper.service.ts
@@ -1,8 +1,9 @@
 import { Injectable } from '@angular/core';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
-import { IdbAnalysisItem, IdbFacility, IdbPredictorEntry, IdbUtilityMeterData } from 'src/app/models/idb';
+import { AnalysisGroup, IdbAnalysisItem, IdbFacility, IdbPredictorEntry, IdbUtilityMeterData } from 'src/app/models/idb';
 import * as _ from 'lodash';
 import { MonthlyData } from 'src/app/models/calanderization';
+import { AnnualAnalysisSummary, MonthlyAnalysisSummaryData } from 'src/app/models/analysis';
 
 @Injectable({
   providedIn: 'root'
@@ -94,5 +95,75 @@ export class AnalysisCalculationsHelperService {
         return meterItemDate >= startDate && meterItemDate < endDate;
       });
     }
+  }
+
+
+  getAnnualAnalysisSummary(modelYear: number, analysisItem: IdbAnalysisItem, facility: IdbFacility, monthlyAnalysisSummaryData: Array<MonthlyAnalysisSummaryData>): Array<AnnualAnalysisSummary>{
+    let annualRegressionSummary: Array<AnnualAnalysisSummary> = new Array();
+    let baselineYear: number = facility.sustainabilityQuestions.energyReductionBaselineYear;
+    let reportYear: number = analysisItem.reportYear;
+    if (facility.fiscalYear == 'nonCalendarYear' && facility.fiscalYearCalendarEnd) {
+      baselineYear = baselineYear - 1;
+      reportYear = reportYear - 1;
+    }
+    let previousYearSavings: number = 0;
+    let previousYearEnergyUse: number = 0;
+    let previousYearModeledEnergyUse: number = 0;
+    let totalEnergySavings: number = 0;
+    let totalModeledEnergySavings: number = 0;
+    let baselineEnergyUse: number;
+    let baselineModeledEnergy: number;
+    let baselineSEnPI: number;
+    for (let summaryYear: number = baselineYear; summaryYear <= reportYear; summaryYear++) {
+      let summaryYearData: Array<MonthlyAnalysisSummaryData> = monthlyAnalysisSummaryData.filter(data => { return data.fiscalYear == summaryYear });
+      let energyUse: number = _.sumBy(summaryYearData, 'totalEnergy');
+      let modeledEnergyUse: number = _.sumBy(summaryYearData, 'modeledEnergy');
+      let SEnPI: number;
+      let cumulativeSavings: number = 0;
+      let annualSavings: number = 0;
+
+      if (summaryYear == baselineYear) {
+        baselineEnergyUse = energyUse;
+        baselineModeledEnergy = modeledEnergyUse;
+      }
+
+      if (summaryYear > modelYear) {
+        SEnPI = (energyUse * baselineModeledEnergy) / (modeledEnergyUse * baselineEnergyUse);
+      } else {
+        SEnPI = modeledEnergyUse / energyUse;
+      }
+      if (summaryYear == baselineYear) {
+        baselineSEnPI = SEnPI;
+        previousYearEnergyUse = energyUse;
+        previousYearModeledEnergyUse = modeledEnergyUse;
+      } else if (summaryYear > modelYear) {
+        cumulativeSavings = 1 - SEnPI;
+        annualSavings = cumulativeSavings - previousYearSavings;
+      } else {
+        cumulativeSavings = (1 - baselineSEnPI) - (1 - SEnPI);
+        annualSavings = cumulativeSavings - previousYearSavings;
+      }
+
+      let annualEnergySavings: number = previousYearEnergyUse - energyUse;
+      let annualModeledEnergySavings: number = previousYearModeledEnergyUse - modeledEnergyUse;
+      totalEnergySavings = totalEnergySavings + annualEnergySavings;
+      totalModeledEnergySavings = totalModeledEnergySavings + annualModeledEnergySavings;
+      annualRegressionSummary.push({
+        year: summaryYear,
+        energyUse: energyUse,
+        annualEnergySavings: annualEnergySavings,
+        totalEnergySavings: totalEnergySavings,
+        annualModeledEnergySavings: annualModeledEnergySavings,
+        totalModeledEnergySavings: totalModeledEnergySavings,
+        modeledEnergyUse: modeledEnergyUse,
+        SEnPI: SEnPI,
+        cumulativeSavings: cumulativeSavings * 100,
+        annualSavings: annualSavings * 100
+      })
+      previousYearSavings = cumulativeSavings;
+      previousYearEnergyUse = energyUse;
+      previousYearModeledEnergyUse = modeledEnergyUse;
+    }
+    return annualRegressionSummary;
   }
 }
