+++ src/app/analysis/run-analysis/group-analysis/absolute-energy-consumption/annual-absolute-energy-consumption/annual-absolute-energy-consumption.component.html
@@ -1,1 +1,15 @@
-<p>annual-absolute-energy-consumption works!</p>
+<div class="controls pb-0 col">
+    <div class="btn-group pull-right">
+        <button class="btn" [ngClass]="{'btn-solid': dataDisplay == 'table'}"
+            (click)="setDataDisplay('table')">Table</button>
+        <button class="btn" [ngClass]="{'btn-solid': dataDisplay == 'graph'}"
+            (click)="setDataDisplay('graph')">Graph</button>
+    </div>
+</div>
+
+<app-annual-analysis-summary-table *ngIf="dataDisplay == 'table'" [annualAnalysisSummary]="annualAnalysisSummary"
+    [analysisItem]="analysisItem">
+</app-annual-analysis-summary-table>
+<!-- <app-annual-regression-analysis-graph *ngIf="dataDisplay == 'graph'" [annualRegressionSummary]="annualRegressionSummary"
+    [analysisItem]="analysisItem">
+</app-annual-regression-analysis-graph> -->
