+++ src/app/analysis/run-analysis/group-analysis/annual-analysis-summary-table/annual-analysis-summary-table.component.css
