+++ src/app/analysis/run-analysis/group-analysis/annual-analysis-summary-table/annual-analysis-summary-table.component.html
@@ -46,7 +46,7 @@
         </tr>
     </thead>
     <tbody>
-        <tr *ngFor="let summary of annualRegressionSummary">
+        <tr *ngFor="let summary of annualAnalysisSummary">
             <td>
                 {{summary.year}}
             </td>
