--- src/app/analysis/run-analysis/group-analysis/regression/annual-regression-analysis/annual-regression-analysis-table/annual-regression-analysis-table.component.ts
@@ -1,20 +1,0 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { AnnualAnalysisSummary } from 'src/app/models/analysis';
-import { IdbAnalysisItem } from 'src/app/models/idb';
-
-@Component({
-  selector: 'app-annual-regression-analysis-table',
-  templateUrl: './annual-regression-analysis-table.component.html',
-  styleUrls: ['./annual-regression-analysis-table.component.css']
-})
-export class AnnualRegressionAnalysisTableComponent implements OnInit {
-  @Input()
-  annualRegressionSummary: Array<AnnualAnalysisSummary>;
-  @Input()
-  analysisItem: IdbAnalysisItem;
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-}
